requirements.txt
# Core FastAPI dependencies
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.6.1

# HTTP client
httpx==0.28.1

# Web scraping
beautifulsoup4==4.12.3
newspaper3k==0.2.8
lxml==5.3.0
lxml_html_clean==0.3.1

# Utilities
python-dateutil==2.9.0.post0
python-multipart==0.0.18

# Optional: Redis caching
redis==5.2.1

# Optional: Monitoring
prometheus-client==0.21.1

# Optional: Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Rate limiting
slowapi==0.1.9

# Testing (optional)
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0